@@ -1,4 +1,4 @@
-module.exports = {
+export default {
 	tags: ["posts"],
 	layout: "layouts/post.njk",
 }
